Improve the Middleware

+++ app/Middleware/SetupRoutePagination.php
@@ -62,7 +62,9 @@ class SetupRoutePagination
 
         Paginator::currentPageResolver(function ($pageName = 'page') use ($route)
         {
-            $page = (int) str_replace($pageName .'/', '', $route->parameter('pageQuery', $pageName .'/1'));
+            $page = str_replace(
+                $pageName .'/', '', $route->parameter('pageQuery', $pageName .'/1')
+            );
 
             if ((filter_var($page, FILTER_VALIDATE_INT) !== false) && ((int) $page >= 1)) {
                 return $page;
